+++ packages/aws-cdk/package.json
@@ -32,11 +32,6 @@
       ]
     }
   },
-  "nyc": {
-    "statements": 8,
-    "lines": 8,
-    "branches": 3
-  },
   "author": {
     "name": "Amazon Web Services",
     "url": "https://aws.amazon.com",
@@ -47,11 +42,11 @@
     "@types/archiver": "^3.0.0",
     "@types/fs-extra": "^8.0.1",
     "@types/glob": "^7.1.1",
+    "@types/jest": "^24.0.25",
     "@types/jszip": "^3.1.6",
     "@types/minimatch": "^3.0.3",
     "@types/mockery": "^1.4.29",
     "@types/node": "^10.17.13",
-    "@types/nodeunit": "^0.0.30",
     "@types/promptly": "^3.0.0",
     "@types/request": "^2.48.4",
     "@types/semver": "^6.2.0",
@@ -62,18 +57,19 @@
     "@types/yargs": "^13.0.4",
     "aws-sdk-mock": "^5.0.0",
     "cdk-build-tools": "1.19.0",
+    "jest": "^24.9.0",
     "jszip": "^3.2.2",
     "mockery": "^2.1.0",
-    "nodeunit": "^0.11.3",
     "pkglint": "1.19.0",
-    "sinon": "^8.0.2"
+    "sinon": "^8.0.2",
+    "ts-jest": "^24.2.0"
   },
   "dependencies": {
     "@aws-cdk/cloudformation-diff": "1.19.0",
     "@aws-cdk/cx-api": "1.19.0",
     "@aws-cdk/region-info": "1.19.0",
     "archiver": "^3.1.1",
-    "aws-sdk": "^2.597.0",
+    "aws-sdk": "^2.596.0",
     "camelcase": "^5.3.1",
     "colors": "^1.4.0",
     "decamelize": "^3.2.0",
@@ -88,7 +84,7 @@
     "table": "^5.4.6",
     "uuid": "^3.3.3",
     "yaml": "^1.7.2",
-    "yargs": "^15.1.0"
+    "yargs": "^15.0.2"
   },
   "repository": {
     "url": "https://github.com/aws/aws-cdk.git",
@@ -103,5 +99,23 @@
   "engines": {
     "node": ">= 10.3.0"
   },
-  "stability": "stable"
+  "stability": "stable",
+  "jest": {
+    "collectCoverage": true,
+    "coverageReporters": [
+      "lcov",
+      "html",
+      "text-summary"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 45,
+        "statements": 60
+      }
+    },
+    "preset": "ts-jest",
+    "testMatch": [
+      "**/?(*.)+(spec|test).ts?(x)"
+    ]
+  }
 }
